backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
aiosqlite==0.20.0
asyncpg==0.31.0
alembic==1.14.1
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
google-generativeai==0.8.5
python-dotenv==1.1.0
boto3==1.38.13
